chore: Change background color and AI behavior
Makes the AI a little more aggressive in its moves and also gives it the ability to sit and wait rather than wobbling near where it needs to be. This makes the AI feel more deliberate in its moves and increases it's chances of catching balls in some circumstances.

The backdrop change makes the CRT filter *much* easier to see.