fix: clone breakout state to new object
Replicant data, including the returned result from useReplicant, is a
recursive proxy object and using internally for breakout state is
causing excessive replicant updates that get out of sync with the
serialization step. This change ensures that the data coming into the
deserialization function is cloned to an unproxied object.