feat: Add Here Comes Niko channel
Adds a Here Comes Niko! themed channel.
Donations trigger the frog to say a cheesy pickup line, just like in the game. The frog can also pull up and play the Tamagotchi!

This is my first try with any CSS or TypeScript coding. I also got explicit permission from the developers at Frog Vibes studios to add this breakscreen, as it is a recent game and it would conflict with the guidelines otherwise.

Checklist:
 I have read and followed the requirements in the Contributing document.
 My code has been tested.
 My commit title follows Conventional Commits formatting.
 All the code is my own, or is code I have the rights to, and is being made available under the Apache License Version 2.0.
 My channel contains no first-party Nintendo assets.
 I understand that GamesDoneQuick may reject, make changes to, or choose not to show this channel on broadcast, and that discussion or merging does not guarantee the work will be used.